RawConfigParser instead of ConfigParser
Since I am now dependent on that the parser do not interpret the conditions it's
own way, and using the 'raw' key-word when getting values from the file, it is
much more reasonable to use this object. Dont have to give the 'raw' key-word
any more.

Plus, I just learned the meaning of RawConfigParser the other day looking
at some old pycon videos.