added todo:s and clarifications and reminders
All in the docstrings. Some stuff more to do but I hold hope. I really want to
get done with this branch so that I can go on with some more fancy stuff like
the idea with the lazyutils module. And a helper pack function for genfromtxt
and the clipboard idea.

I am using the commit messages as some sort of diary and it feels nice. Helps to
remember trails of thoughts.

Anyway, I will finish with this evolve branch, merge it into master and release.
Or, first I guess I'll have to fix the documentation. OK, I fix this branch,
make a new branch of it for the docs, merge it back to this branch and then
finally merge into master branch. That's a plan.