None instead of nan if not number
And a bunch of docstring edits. The important part of the commit is in datautils
module. Not doing an explicit cast to object, but using a None array if the
input array is not numberlike.

For some reason I am hesitant to do the cast myself when it seems numpy does
this (as expected?) when the 'spare' array is full of None's.

The change is the masked function.