nof is corrected and more doctests and something that fails
And it fails because of interesting discovery. The load method is not that
smooth as I thought. One has to re-put the arguments as they were put last time,
and that feels wrong when there is such a function as this load thing. It should
be smooth.

It might be possible to store away the args and kwargs in the instance so load
can be called like this: load(inst.args, inst.kwargs).

Next, the loadfunc used for sheetpack and the sheetpack convenience function
must be re-written. Now its really messed up. The loadfunc itself is taking both
the file name and a loaded xlrd sheet. That cannot be.