reload function and change of sheetpack function
The intention is to have the pack re-loadable no matter how the pack was
created. Of course the new data set has to be of the same structure.

Later I will test what happens when it is not of the same structure, I mean for
example if the number of channels is different in the new file.

There is also some clean-ups in this commit. Code that I considered wrong.