check on the keys in chnames in load and delete the reload method
The idea with reload I got when there was a problem using load on a running
instance from sheetpack function. When that problem was fixed there is really no
problem. Just repeat the arguments. the idea with keeping the args and kwargs is
interesting but I discard it for this time.