lots of doctests lots of additions and work with ChannelPack
Could not resist to add a counter method. I hope it will not feel out of scope.
Also a parts method is added for easy iteration over the parts.

I am now getting closer to a merge into master. Need some more testing and
tuning. This branch is a big work.